refactored components folder to remove unneeded nesting
