+++ src/components/EventCardThumbnail.jsx
