+++ src/components/Layout.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { motion } from "framer-motion";
-import "../../styles/index.scss";
-import { Navbar, Footer } from "../";
-import routes from "../../utils/routes";
+import "../styles/index.scss";
+import { Navbar, Footer } from ".";
+import routes from "../utils/routes";
 
 const Layout = ({ children }) => {
   return (
