+++ src/components/SlideShow.jsx
