+++ src/images/svg/hesperosLogoPathData.js
