+++ src/images/svg/socialMediaPathData.js
